+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/mixin/MobMixin.java
@@ -25,6 +25,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class MobMixin implements MobMixinAddon {
 
     private static final EntityDataAccessor<Float> SCALING$RANDOM_MOB_SIZES = SynchedEntityData.defineId(Mob.class, EntityDataSerializers.FLOAT);
+    private static final EntityDataAccessor<Boolean> SCALE_LOOT$RANDOM_MOB_SIZES = SynchedEntityData.defineId(Mob.class, EntityDataSerializers.BOOLEAN);
+    private static final EntityDataAccessor<Boolean> SCALE_XP$RANDOM_MOB_SIZES = SynchedEntityData.defineId(Mob.class, EntityDataSerializers.BOOLEAN);
 
     @Override
     public float getMobScaling$RandomMobSizes() {
@@ -39,6 +41,26 @@ public abstract class MobMixin implements MobMixinAddon {
     @Override
     public EntityDataAccessor<Float> getTracker$RandomMobSizes() {
         return SCALING$RANDOM_MOB_SIZES;
+    }
+
+    @Override
+    public boolean shouldScaleLoot$RandomMobSizes() {
+        return ((Mob) (Object) this).getEntityData().get(SCALE_LOOT$RANDOM_MOB_SIZES);
+    }
+
+    @Override
+    public void setShouldScaleLoot$RandomMobSizes(boolean shouldScale) {
+        ((Mob) (Object) this).getEntityData().set(SCALE_LOOT$RANDOM_MOB_SIZES, shouldScale);
+    }
+
+    @Override
+    public boolean shouldScaleXP$RandomMobSizes() {
+        return ((Mob) (Object) this).getEntityData().get(SCALE_XP$RANDOM_MOB_SIZES);
+    }
+
+    @Override
+    public void setShouldScaleXP$RandomMobSizes(boolean shouldScale) {
+        ((Mob) (Object) this).getEntityData().set(SCALE_XP$RANDOM_MOB_SIZES, shouldScale);
     }
 
     @Inject(at = @At("TAIL"), method = "finalizeSpawn")
@@ -65,6 +87,10 @@ public abstract class MobMixin implements MobMixinAddon {
                 this.addModifier$RandomMobSizes(Attributes.MOVEMENT_SPEED, speedScaling, false);
             }
         }
+        boolean shouldScaleLoot = sampler == null || sampler.shouldScaleLoot();
+        this.setShouldScaleLoot$RandomMobSizes(shouldScaleLoot);
+        boolean shouldScaleXP = sampler == null || sampler.shouldScaleXP();
+        this.setShouldScaleXP$RandomMobSizes(shouldScaleXP);
         this.setMobScaling$RandomMobSizes(scaling);
     }
 
@@ -81,6 +107,8 @@ public abstract class MobMixin implements MobMixinAddon {
     @Inject(at = @At("TAIL"), method = "defineSynchedData")
     private void defineSyncedData$RandomMobSizes(CallbackInfo ci) {
         ((Mob) (Object) this).getEntityData().define(SCALING$RANDOM_MOB_SIZES, 1.0F);
+        ((Mob) (Object) this).getEntityData().define(SCALE_LOOT$RANDOM_MOB_SIZES, true);
+        ((Mob) (Object) this).getEntityData().define(SCALE_XP$RANDOM_MOB_SIZES, true);
     }
 
     @Inject(at = @At("TAIL"), method = "readAdditionalSaveData")
@@ -88,11 +116,19 @@ public abstract class MobMixin implements MobMixinAddon {
         if (tag.contains("ScaleFactor")) {
             this.setMobScaling$RandomMobSizes(tag.getFloat("ScaleFactor"));
         }
+        if (tag.contains("ScaleLoot")) {
+            this.setShouldScaleLoot$RandomMobSizes(tag.getBoolean("ScaleLoot"));
+        }
+        if (tag.contains("ScaleExperience")) {
+            this.setShouldScaleXP$RandomMobSizes(tag.getBoolean("ScaleExperience"));
+        }
     }
 
     @Inject(at = @At("TAIL"), method = "addAdditionalSaveData")
     private void addAdditionalSaveData$RandomMobSizes(CompoundTag tag, CallbackInfo ci) {
         tag.putFloat("ScaleFactor", this.getMobScaling$RandomMobSizes());
+        tag.putBoolean("ScaleLoot", this.shouldScaleLoot$RandomMobSizes());
+        tag.putBoolean("ScaleExperience", this.shouldScaleXP$RandomMobSizes());
     }
 
     @Inject(at = @At("TAIL"), method = "convertTo")
@@ -101,8 +137,12 @@ public abstract class MobMixin implements MobMixinAddon {
             return;
         }
         float scaling = ((MobMixinAddon) this).getMobScaling$RandomMobSizes();
+        boolean scaleLoot = ((MobMixinAddon) this).shouldScaleLoot$RandomMobSizes();
+        boolean scaleXP = ((MobMixinAddon) this).shouldScaleXP$RandomMobSizes();
         T entity = cir.getReturnValue();
         ((MobMixinAddon) entity).setMobScaling$RandomMobSizes(scaling);
+        ((MobMixinAddon) entity).setShouldScaleLoot$RandomMobSizes(scaleLoot);
+        ((MobMixinAddon) entity).setShouldScaleXP$RandomMobSizes(scaleXP);
         RandomMobSizes.LOGGER.info("Converted '{}' to '{}' with scaling '{}'", this.getClass().getSimpleName(), entity.getClass().getSimpleName(), scaling);
     }