+++ Common/src/main/java/com/tristankechlo/random_mob_sizes/sampler/ScalingSampler.java
@@ -3,6 +3,7 @@ package com.tristankechlo.random_mob_sizes.sampler;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
+import com.tristankechlo.random_mob_sizes.RandomMobSizes;
 import com.tristankechlo.random_mob_sizes.sampler.types.StaticScalingSampler;
 import net.minecraft.util.GsonHelper;
 import net.minecraft.util.Mth;
@@ -17,6 +18,8 @@ public abstract class ScalingSampler {
     protected final AttributeScalingTypes scaleDamage;
     protected final AttributeScalingTypes scaleSpeed;
     protected boolean addAttributeModifiers = true;
+    protected boolean shouldScaleLoot = true;
+    protected boolean shouldScaleXP = true;
 
     protected ScalingSampler() {
         this(AttributeScalingTypes.NORMAL, AttributeScalingTypes.NORMAL, AttributeScalingTypes.INVERSE);
@@ -33,6 +36,8 @@ public abstract class ScalingSampler {
         this.scaleHealth = AttributeScalingTypes.byName(GsonHelper.getAsString(json, "scale_health", "none"));
         this.scaleDamage = AttributeScalingTypes.byName(GsonHelper.getAsString(json, "scale_damage", "none"));
         this.scaleSpeed = AttributeScalingTypes.byName(GsonHelper.getAsString(json, "scale_speed", "none"));
+        this.shouldScaleLoot = getBooleanSafe(json, "scale_loot", entityType, true);
+        this.shouldScaleXP = getBooleanSafe(json, "scale_xp", entityType, true);
     }
 
     protected abstract float sampleScalingFactor(RandomSource random, Difficulty difficulty);
@@ -48,6 +53,8 @@ public abstract class ScalingSampler {
         JsonObject json = new JsonObject();
         json = serialize(json);
         if (this.addAttributeModifiers) {
+            json.addProperty("scale_loot", this.shouldScaleLoot);
+            json.addProperty("scale_xp", this.shouldScaleXP);
             json.addProperty("scale_health", scaleHealth.getSerializedName());
             json.addProperty("scale_damage", scaleDamage.getSerializedName());
             json.addProperty("scale_speed", scaleSpeed.getSerializedName());
@@ -77,6 +84,14 @@ public abstract class ScalingSampler {
 
     public AttributeScalingTypes getSpeedScaler() {
         return scaleSpeed;
+    }
+
+    public boolean shouldScaleLoot() {
+        return shouldScaleLoot;
+    }
+
+    public boolean shouldScaleXP() {
+        return shouldScaleXP;
     }
 
     public static ScalingSampler deserializeSampler(JsonElement jsonElement, String entityType) {
@@ -116,6 +131,16 @@ public abstract class ScalingSampler {
 
     protected static boolean isFloatOutOfBounds(float value, float min, float max) {
         return value < min || value > max;
+    }
+
+    private boolean getBooleanSafe(JsonObject json, String key, String entityType, boolean def) {
+        if (GsonHelper.isBooleanValue(json, key)) {
+            return GsonHelper.getAsBoolean(json, key);
+        } else {
+            RandomMobSizes.LOGGER.error("Expected '{}' to be a Boolean", key);
+            RandomMobSizes.LOGGER.error("Error while parsing config value '{}' for entity '{}', using default value.", key, entityType);
+        }
+        return def;
     }
 
 }